+++ app-backend/backsplash-server/src/main/scala/com/rasterfoundry/backsplash/services/MosaicService.scala
@@ -4,10 +4,14 @@ import com.rasterfoundry.common.datamodel.User
 import com.rasterfoundry.backsplash._
 import com.rasterfoundry.backsplash.Parameters._
 import com.rasterfoundry.common.utils.TileUtils
+import com.rasterfoundry.database.ProjectLayerDao
 
 import cats.data.Validated._
 import cats.effect.{ContextShift, IO}
 import cats.implicits._
+import com.azavea.maml.eval.BufferingInterpreter
+import com.azavea.maml.ast.{GeomLit, Masking, RasterVar}
+import doobie.implicits._
 import geotrellis.proj4.{LatLng, WebMercator}
 import geotrellis.raster.io.geotiff.MultibandGeoTiff
 import geotrellis.server._
@@ -19,7 +23,8 @@ import org.http4s.dsl.io._
 import org.http4s.headers._
 import org.http4s.util.CaseInsensitiveString
 import doobie.util.transactor.Transactor
-import geotrellis.vector.{Polygon, Projected}
+import geotrellis.vector.{MultiPolygon, Polygon, Projected}
+import geotrellis.vector.io._
 
 import java.util.UUID
 
@@ -45,16 +50,34 @@ class MosaicService[LayerStore: ProjectStore, HistStore, ToolStore](
             bandOverride) as user =>
         val polygonBbox: Projected[Polygon] =
           TileUtils.getTileBounds(z, x, y)
-        val eval = LayerTms.identity(
-          layers.read(layerId, Some(polygonBbox), bandOverride, None)
-        )
+        val getEval = ProjectLayerDao
+          .unsafeGetProjectLayerById(layerId)
+          .transact(xa) map { layer =>
+          layer.geometry flatMap { _.geom.as[MultiPolygon] } match {
+            case Some(mask) =>
+              // Intermediate val to anchor the implicit resolution with multiple argument lists
+              val expression =
+                Masking(
+                  List(GeomLit(mask.toGeoJson), RasterVar("mosaic"))
+                )
+              val param =
+                layers.read(layerId, Some(polygonBbox), bandOverride, None)
+              LayerTms(IO.pure(expression),
+                       IO.pure(Map("mosaic" -> param)),
+                       BufferingInterpreter.DEFAULT)
+            case _ =>
+              LayerTms.identity(
+                layers.read(layerId, Some(polygonBbox), bandOverride, None)
+              )
+          }
+        }
 
         for {
           fiberAuthProject <- authorizers.authProject(user, projectId).start
           fiberAuthLayer <- authorizers
             .authProjectLayer(projectId, layerId)
             .start
-          fiberResp <- eval(z, x, y).start
+          fiberResp <- getEval.flatMap(_(z, x, y)).start
           _ <- (fiberAuthProject, fiberAuthLayer).tupled.join
             .handleErrorWith { error =>
               fiberResp.cancel *> IO.raiseError(error)